#243 Upgrade chart configuration
