+++ monitor-ui/src/components/pages/DIY-view/view-config.vue
@@ -70,7 +70,7 @@
 <script>
 import {generateUuid} from '@/assets/js/utils'
 import VueGridLayout from 'vue-grid-layout'
-import {drawChart} from  '@/assets/config/chart-rely'
+import {readyToDraw} from  '@/assets/config/chart-rely'
 const echarts = require('echarts/lib/echarts');
 export default {
   name: '',
@@ -113,51 +113,57 @@ export default {
         })) 
       })
       this.$root.$httpRequestEntrance.httpRequestEntrance('GET',this.$root.apiCenter.metricConfigView.api, {config: `[${params.join(',')}]`}, responseData => {
-        var legend = []
-        if (responseData.series.length === 0) {
-          this.noDataTip = true
-          return
-        }
-        const colorX = ['#339933','#33CCCC','#666699','#66CC66','#996633','#9999CC','#339966','#663333','#6666CC','#336699','#3399CC','#33CC66','#CC3333','#CC6666','#996699','#CC9933']
-        let colorSet = []
-        for (let i=0;i<colorX.length;i++) {
-          let tmpIndex = viewIndex*3 + i
-          tmpIndex = tmpIndex%colorX.length
-          colorSet.push(colorX[tmpIndex])
-        }
-        responseData.series.forEach((item,index)=>{
-          legend.push(item.name)
-          item.symbol = 'none'
-          item.smooth = true
-          item.lineStyle = {
-            width: 1
-          }
-          item.itemStyle = {
-            normal:{
-              color: colorSet[index] === undefined ? '#666699': colorSet[index]
-            }
-          }
-          item.areaStyle = {
-            normal: {
-              color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [{
-                  offset: 0,
-                  color: colorSet[index] === undefined ? '#666699': colorSet[index]
-              }, {
-                  offset: 1,
-                  color: 'white'
-              }])
-            }
-          }
-        }) 
+        // var legend = []
+        // if (responseData.series.length === 0) {
+        //   this.noDataTip = true
+        //   return
+        // }
+        // const colorX = ['#339933','#33CCCC','#666699','#66CC66','#996633','#9999CC','#339966','#663333','#6666CC','#336699','#3399CC','#33CC66','#CC3333','#CC6666','#996699','#CC9933']
+        // let colorSet = []
+        // for (let i=0;i<colorX.length;i++) {
+        //   let tmpIndex = viewIndex*3 + i
+        //   tmpIndex = tmpIndex%colorX.length
+        //   colorSet.push(colorX[tmpIndex])
+        // }
+        // responseData.series.forEach((item,index)=>{
+        //   legend.push(item.name)
+        //   item.symbol = 'none'
+        //   item.smooth = true
+        //   item.lineStyle = {
+        //     width: 1
+        //   }
+        //   item.itemStyle = {
+        //     normal:{
+        //       color: colorSet[index] === undefined ? '#666699': colorSet[index]
+        //     }
+        //   }
+        //   item.areaStyle = {
+        //     normal: {
+        //       color: new echarts.graphic.LinearGradient(0, 0, 0, 1, [{
+        //           offset: 0,
+        //           color: colorSet[index] === undefined ? '#666699': colorSet[index]
+        //       }, {
+        //           offset: 1,
+        //           color: 'white'
+        //       }])
+        //     }
+        //   }
+        // }) 
+        // responseData.yaxis.unit =  panalUnit  
+        // let config = {
+        //   title: responseData.title,
+        //   legend: legend,
+        //   series: responseData.series,
+        //   yaxis: responseData.yaxis
+        // }
+        // this.elId = id
+        // drawChart(this, config, {eye: false,dataZoom:false})
+
         responseData.yaxis.unit =  panalUnit  
-        let config = {
-          title: responseData.title,
-          legend: legend,
-          series: responseData.series,
-          yaxis: responseData.yaxis
-        }
         this.elId = id
-        drawChart(this, config, {eye: false,dataZoom:false})
+        const chartConfig = {eye: false,dataZoom:false}
+
+        readyToDraw(this,responseData, viewIndex, chartConfig)
       })
     },
     addItem() {
